requirements.txt
email_validator
Flask
pyodbc
wtforms
flask-wtf 
flask-login
flask-sqlalchemy
pylint
pylint-flask
sqlalchemy-datatables
pandas
Flask-Caching